refactor(geode-core): group driven solver + ports/extraction/scattering under driven
Move the frequency-domain driven solver and its satellites into a single
directory-backed `driven` module, mirroring the merged `eigen` layout:

  driven/{mod.rs, solve.rs, ports/{mod.rs, lumped.rs, wave.rs},
          extraction.rs, scattering.rs}

- git mv driven.rs -> driven/solve.rs, lumped_port.rs/wave_port.rs ->
  driven/ports/{lumped,wave}.rs, extraction.rs/scattering.rs ->
  driven/{extraction,scattering}.rs (history preserved).
- driven/mod.rs declares solve/ports/extraction/scattering without
  re-exporting up a level; ports/mod.rs flattens the private lumped/wave
  leaves into one driven::ports namespace.
- lib.rs: drop the four obsolete module decls; repoint the five flat
  re-export blocks at canonical paths as #[deprecated] shims (public
  items only). The four pub(crate) scattering helpers (BoxFaceSample,
  box_surface_samples, cross, tet_geometry) relocate with no shim;
  ntff.rs imports repointed to crate::driven::scattering.
- prelude.rs: add high-traffic driven items from canonical paths.
- Migrate all Class A src use/path sites and all Class B intra-doc links
  to the new driven::{solve,ports,extraction,scattering} paths.

silvermuller_self_consistent.rs stays top-level (eigen follow-up).
Pure refactor, no behavior change. Class C/D test/bench/example item
uses deferred to #387.
